Bugs y mensajes inglés

+++ src/category/category.service.ts
@@ -3,16 +3,20 @@ import { CreateUpdateCategoryDto } from './dto/createUpdate-category.dto';
 import { ICategoryRepository } from './repositories/category.repository.interface';
 import { Users } from '@/users/entities/users.entity';
 import { Category } from './entities/category.entity';
+import { IUsersRepository } from '@/users/repositories/users.repository.interface';
 
 
 @Injectable()
 export class CategoryService {
   constructor(
-      @Inject('ICategoryRepository') // 👈 Inyecta la interfaz
+      @Inject('ICategoryRepository') // 👈 Inyecta las interfaces
       private readonly categoryRepository: ICategoryRepository,
+      @Inject('IUsersRepository') 
+      private readonly userRepository: IUsersRepository,
     ) {}
      
-  getByUser(userId: number) {
+  async getByUser(userId: number) {
+    await this.checkUserPermission(userId)
     return this.categoryRepository.getByUser(userId)
   }
 
@@ -21,6 +25,7 @@ export class CategoryService {
   }
 
   async create(userId: number, createUpdateCategoryDto: CreateUpdateCategoryDto): Promise<Category> {
+    await this.checkUserPermission(userId)
     return this.categoryRepository.create({
       name: createUpdateCategoryDto.name,
       user: { id: userId }
@@ -28,6 +33,7 @@ export class CategoryService {
   }
 
   async update(userId: number, id: number, CreateUpdateCategoryDto: CreateUpdateCategoryDto) {
+    await this.checkUserPermission(userId)
     await this.checkPermission(userId, id)
     return this.categoryRepository.update(id, {
       name: CreateUpdateCategoryDto.name
@@ -35,10 +41,18 @@ export class CategoryService {
   }
 
   async delete(userId: number, id: number) {
+    await this.checkUserPermission(userId)
     await this.checkPermission(userId, id)
     return this.categoryRepository.delete(id)
   }
-
+  private async checkUserPermission(userId: number){
+    const user = await this.userRepository.getById(userId)
+    console.log("se está llegando")
+    console.log(user)
+    if(!user){
+      throw new ForbiddenException('Your account has been deleted and you do not have permission')
+    }
+  }
   private async checkPermission(userId: number, categoryId: number): Promise<Category> {
     const category = await this.categoryRepository.getById(categoryId);
